+++ packages/ccd-js-gen/jest.config.ts
@@ -1,13 +1,14 @@
 import type { Config } from 'jest';
-
-const esModules = ['@noble/ed25519'].join('|');
+import type {} from 'ts-jest';
 
 const config: Config = {
     preset: 'ts-jest/presets/js-with-ts-esm',
     moduleNameMapper: {
         '^(\\.\\.?\\/.+)\\.js$': '$1', // Remap esmodule file extensions
     },
-    transformIgnorePatterns: [`node_modules/(?!${esModules})`],
+    transformIgnorePatterns: [
+        'node_modules/(?!@noble/ed25519)', // @noble/ed25519 is an ES module only
+    ],
 };
 
 export default config;