+++ packages/sdk/src/types/ContractName.ts
@@ -10,7 +10,7 @@ import {
  * The {@linkcode TypedJsonDiscriminator} discriminator associated with {@linkcode Type} type.
  */
 export const JSON_DISCRIMINATOR = TypedJsonDiscriminator.ContractName;
-type Serializable = string;
+export type Serializable = string;
 
 /** The name of a smart contract. Note: This does _not_ including the 'init_' prefix. */
 class ContractName {
@@ -88,18 +88,27 @@ export function toString(contractName: ContractName): string {
     return contractName.value;
 }
 
-/** Type used when encoding a contract name for the schema. */
+/** Type used when encoding a contract name in the JSON format used when serializing using a smart contract schema type. */
 export type SchemaValue = {
     contract: string;
 };
 
 /**
- * Get contract name in the format used by schema.
+ * Get contract name in the JSON format used when serializing using a smart contract schema type.
  * @param {ContractName} contractName The contract name.
- * @returns {SchemaValue} The schema value representation.
+ * @returns {SchemaValue} The schema JSON representation.
  */
 export function toSchemaValue(contractName: ContractName): SchemaValue {
     return { contract: contractName.value };
+}
+
+/**
+ * Convert to contract name from JSON format used when serializing using a smart contract schema type.
+ * @param {SchemaValue} contractName The contract name in schema JSON format.
+ * @returns {ContractName} The contract name.
+ */
+export function fromSchemaValue(contractName: SchemaValue): ContractName {
+    return fromString(contractName.contract);
 }
 
 /**
@@ -118,10 +127,10 @@ export function equals(left: ContractName, right: ContractName): boolean {
  * @param {Type} value - The account address instance to transform.
  * @returns {TypedJson} The transformed object.
  */
-export function toTypedJSON({ value }: ContractName): TypedJson<Serializable> {
+export function toTypedJSON(value: ContractName): TypedJson<Serializable> {
     return {
         ['@type']: JSON_DISCRIMINATOR,
-        value,
+        value: toString(value),
     };
 }