+++ packages/sdk/src/types/ReceiveName.ts
@@ -12,7 +12,7 @@ import {
  * The {@linkcode TypedJsonDiscriminator} discriminator associated with {@linkcode Type} type.
  */
 export const JSON_DISCRIMINATOR = TypedJsonDiscriminator.ReceiveName;
-type Serializable = string;
+export type Serializable = string;
 
 /**
  * Represents a receive-function in a smart contract module.
@@ -135,21 +135,30 @@ export function toEntrypointName(
     return EntrypointName.fromStringUnchecked(entrypointName);
 }
 
-/** Type used when encoding a receive-name using a schema. */
+/** Type used when encoding a receive-name in the JSON format used when serializing using a smart contract schema type. */
 export type SchemaValue = {
     contract: string;
     func: string;
 };
 
 /**
- * Get receiveName in the format used by schemas.
+ * Get receiveName in the JSON format used when serializing using a smart contract schema type.
  * @param {ReceiveName} receiveName The receive name.
- * @returns {SchemaValue} The schema value representation.
+ * @returns {SchemaValue} The schema JSON representation.
  */
 export function toSchemaValue(receiveName: ReceiveName): SchemaValue {
     const contract = ContractName.toString(toContractName(receiveName));
     const func = EntrypointName.toString(toEntrypointName(receiveName));
     return { contract, func };
+}
+
+/**
+ * Convert to smart contract receive name from JSON format used when serializing using a smart contract schema type.
+ * @param {SchemaValue} receiveName The receive name in schema JSON format.
+ * @returns {ReceiveName} The receive name.
+ */
+export function fromSchemaValue(receiveName: SchemaValue): ReceiveName {
+    return fromString(`${receiveName.contract}.${receiveName.func}`);
 }
 
 /**