+++ packages/sdk/src/types/Timestamp.ts
@@ -9,7 +9,7 @@ import {
  * The {@linkcode TypedJsonDiscriminator} discriminator associated with {@linkcode Type} type.
  */
 export const JSON_DISCRIMINATOR = TypedJsonDiscriminator.Timestamp;
-type Serializable = string;
+export type Serializable = string;
 
 /** Represents a timestamp. */
 class Timestamp {
@@ -59,16 +59,25 @@ export function fromDate(date: Date): Timestamp {
     return fromMillis(date.getTime());
 }
 
-/** Type used when encoding the account address using a schema. */
+/** Type used when encoding a timestamp in the JSON format used when serializing using a smart contract schema type. */
 export type SchemaValue = string;
 
 /**
- * Get timestamp in the format used by schemas.
+ * Get timestamp in the JSON format used when serializing using a smart contract schema type.
  * @param {Timestamp} timestamp The timestamp.
  * @returns {SchemaValue} The schema value representation.
  */
 export function toSchemaValue(timestamp: Timestamp): SchemaValue {
     return toDate(timestamp).toISOString();
+}
+
+/**
+ * Convert to timestamp from JSON format used when serializing using a smart contract schema type.
+ * @param {SchemaValue} timestamp The timestamp in schema format.
+ * @returns {Timestamp} The timestamp
+ */
+export function fromSchemaValue(timestamp: SchemaValue): Timestamp {
+    return fromMillis(Date.parse(timestamp));
 }
 
 /**
@@ -104,7 +113,23 @@ export function toProto(timestamp: Timestamp): Proto.Timestamp {
     };
 }
 
-const fromSerializable = (v: Serializable) => fromMillis(BigInt(v));
+/**
+ * Constructs a {@linkcode Type} from {@linkcode Serializable}.
+ * @param {Serializable} value
+ * @returns {Type} The duration.
+ */
+export function fromSerializable(value: Serializable): Type {
+    return fromMillis(BigInt(value));
+}
+
+/**
+ * Converts {@linkcode Type} into {@linkcode Serializable}
+ * @param {Type} value
+ * @returns {Serializable} The serializable value
+ */
+export function toSerializable(value: Type): Serializable {
+    return value.value.toString();
+}
 
 /**
  * Takes an {@linkcode Type} and transforms it to a {@linkcode TypedJson} format.
@@ -112,10 +137,10 @@ const fromSerializable = (v: Serializable) => fromMillis(BigInt(v));
  * @param {Type} value - The account address instance to transform.
  * @returns {TypedJson} The transformed object.
  */
-export function toTypedJSON({ value }: Timestamp): TypedJson<Serializable> {
+export function toTypedJSON(value: Timestamp): TypedJson<Serializable> {
     return {
         ['@type']: JSON_DISCRIMINATOR,
-        value: value.toString(),
+        value: toSerializable(value),
     };
 }